backend/python-processor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.4
supabase==2.0.2
python-dotenv==1.0.0
pydantic==2.5.0
openpyxl==3.1.2
